Daniel Pittman <daniel@danann.net>:
* Update comment regarding EFS and RCP in the header.

* Fixed a bug with tilde-expansion and remote commands.

The function 'shell-quote-argument' was used to produce a safe path for
remote use in a number of places. This had the unfortunate side effect
of running commands like 'test -f \~/test.file' and failing because the
tilde was not expanded.

We now call 'rcp-handle-expand-file-name' for all paths that will be
shell-quoted later on. It's a touch more costly than I like but, well,
it stops the lisp '(rcp-handle-file-attributes "/r:machine:~/existing.file")'
from failing in a messy way.

* Removed your old comment regarding 'vc-user-login-name'

* Added advice and defun for 'vc-file-owner'

This removes any calls to 'vc-user-login-name' with a parameter, at
least for RCP files. This should (in theory) fix the need to hack at
that function.

* My fix for the *long* delay on each action introduced in
  'rcp-wait-for-output'

There was a one second (!) wait for input before trying to force some.
This is, I guess, a bug fix somewhere else. Unfortunately, one second
for every remote operation (or close to that) is painful.

This makes it wait ~10ms for input, a much more reasonable time for
something intended to timeout mostly. This makes RCP usable again for
me.