* Makefile.in (savannah): Don't call make for texi.  Only the trunk should place its files there.
* lisp/tramp.el (top): Apply `def-edebug-spec' only if function is
defined.  This is not the case for XEmacs without package
"edebug".
(tramp-set-auto-save-file-modes): Set permissions of autosaved
remote files to the permissions of the original file.  This is not
the case for Emacs < 21.3.50 and XEmacs < 21.5.  Add function to
`auto-save-hook'.  Reported by Thomas Prokosch <thomas@nadev.net>.

* texi/Makefile.in (savannah): Copy also index files.

* texi/tramp.texi (Frequently Asked Questions): Comment paragraph about
plink link.  The URL is outdated.  Originator contacted for
clarification.